+++ QuanLyDuLichVaDatTour/Business/DTOs/Tour/SearchTourRequestDto.cs
@@ -4,9 +4,9 @@ public class SearchTourRequestDto
 {
     public string? Keyword { get; set; }
 
-    public ulong? DiemXuatPhatId { get; set; }
+    public long? DiemXuatPhatId { get; set; }
 
-    public List<ulong>? LoaiTourIds { get; set; }
+    public List<long>? LoaiTourIds { get; set; }
 
     public List<string>? PhuongTiens { get; set; }
 
@@ -14,7 +14,7 @@ public class SearchTourRequestDto
 
     public decimal? MaxPrice { get; set; }
 
-    public byte? MinSoNgay { get; set; }
+    public int? MinSoNgay { get; set; }
 
-    public byte? MaxSoNgay { get; set; }
+    public int? MaxSoNgay { get; set; }
 }
