+++ QuanLyDuLichVaDatTour/Business/DTOs/Tour/TourAdminResponseDto.cs
@@ -2,23 +2,23 @@ namespace BLL.DTOs.Tour;
 
 public class TourAdminResponseDto
 {
-    public ulong Id { get; set; }
+    public long Id { get; set; }
 
     public string MaTour { get; set; } = string.Empty;
 
     public string TenTour { get; set; } = string.Empty;
 
-    public ulong LoaiTourId { get; set; }
+    public long LoaiTourId { get; set; }
 
     public string TenLoaiTour { get; set; } = string.Empty;
 
-    public ulong DiemXuatPhatId { get; set; }
+    public long DiemXuatPhatId { get; set; }
 
     public string TenDiemXuatPhat { get; set; } = string.Empty;
 
-    public byte SoNgay { get; set; }
+    public int SoNgay { get; set; }
 
-    public byte SoDem { get; set; }
+    public int SoDem { get; set; }
 
     public string? PhuongTien { get; set; }
